us/pyproject.toml
[tool.poetry]
name = "us"
version = "0.1.0"
description = ""
authors = ["Stuart Lynn <stuart.lynn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.8.5"
aiosignal = "1.3.1"
alembic = "1.12.0"
annotated-types = "0.5.0"
anyio = "4.0.0"
apache-airflow = "2.7.1"
apache-airflow-providers-common-sql = "1.7.1"
apache-airflow-providers-ftp = "3.5.1"
apache-airflow-providers-http = "4.5.1"
apache-airflow-providers-imap = "3.3.1"
apache-airflow-providers-sqlite = "3.4.3"
apispec = "6.3.0"
argcomplete = "3.1.1"
asgiref = "3.7.2"
async-timeout = "4.0.3"
attrs = "23.1.0"
babel = "2.12.1"
backoff = "1.10.0"
blinker = "1.6.2"
cachelib = "0.9.0"
cattrs = "23.1.2"
certifi = "2023.7.22"
cffi = "1.15.1"
charset-normalizer = "3.2.0"
click = "8.1.7"
clickclick = "20.10.2"
colorama = "0.4.6"
colorlog = "4.8.0"
configupdater = "3.1.1"
connexion = "2.14.2"
cron-descriptor = "1.4.0"
croniter = "1.4.1"
cryptography = "41.0.3"
deprecated = "1.2.14"
dill = "0.3.1.1"
dnspython = "2.4.2"
docutils = "0.20.1"
email-validator = "1.3.1"
flask = "2.2.5"
flask-appbuilder = "4.3.6"
flask-babel = "2.0.0"
flask-caching = "2.0.2"
flask-jwt-extended = "4.5.2"
flask-limiter = "3.5.0"
flask-login = "0.6.2"
flask-session = "0.5.0"
flask-sqlalchemy = "2.5.1"
flask-wtf = "1.1.1"
frozenlist = "1.4.0"
google-re2 = "1.1"
googleapis-common-protos = "1.60.0"
graphviz = "0.20.1"
grpcio = "1.57.0"
gunicorn = "21.2.0"
h11 = "0.14.0"
httpcore = "0.16.3"
httpx = "0.23.3"
idna = "3.4"
importlib-metadata = "6.8.0"
importlib-resources = "6.0.1"
inflection = "0.5.1"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
jsonschema = "4.19.0"
jsonschema-specifications = "2023.7.1"
lazy-object-proxy = "1.9.0"
limits = "3.6.0"
linkify-it-py = "2.0.2"
lockfile = "0.12.2"
mako = "1.2.4"
markdown = "3.4.4"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.20.1"
marshmallow-oneofschema = "3.0.1"
marshmallow-sqlalchemy = "0.26.1"
mdit-py-plugins = "0.4.0"
mdurl = "0.1.2"
multidict = "6.0.4"
opentelemetry-api = "1.20.0"
opentelemetry-exporter-otlp = "1.20.0"
opentelemetry-exporter-otlp-proto-common = "1.20.0"
opentelemetry-exporter-otlp-proto-grpc = "1.20.0"
opentelemetry-exporter-otlp-proto-http = "1.20.0"
opentelemetry-proto = "1.20.0"
opentelemetry-sdk = "1.20.0"
opentelemetry-semantic-conventions = "0.41b0"
ordered-set = "4.1.0"
packaging = "23.1"
pathspec = "0.11.2"
pendulum = "2.1.2"
pluggy = "1.3.0"
prison = "0.2.1"
protobuf = "4.21.12"
psutil = "5.9.5"
pycparser = "2.21"
pydantic = "2.3.0"
pydantic-core = "2.6.3"
pygments = "2.16.1"
pyjwt = "2.8.0"
python-daemon = "3.0.1"
python-dateutil = "2.8.2"
python-nvd3 = "0.15.0"
python-slugify = "8.0.1"
pytz = "2023.3"
pytzdata = "2020.1"
pyyaml = "6.0.1"
referencing = "0.30.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rfc3339-validator = "0.1.4"
rfc3986 = "1.5.0"
rich = "13.5.2"
rich-argparse = "1.3.0"
rpds-py = "0.10.2"
setproctitle = "1.3.2"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "1.4.49"
sqlalchemy-jsonfield = "1.0.1.post0"
sqlalchemy-utils = "0.41.1"
sqlparse = "0.4.4"
tabulate = "0.9.0"
tenacity = "8.2.3"
termcolor = "2.3.0"
text-unidecode = "1.3"
typing-extensions = "4.7.1"
uc-micro-py = "1.0.2"
unicodecsv = "0.14.1"
urllib3 = "1.26.16"
werkzeug = "2.2.3"
wrapt = "1.15.0"
wtforms = "3.0.1"
yarl = "1.9.2"
zipp = "3.16.2"
python-lsp-server = "^1.8.0"
polars = "^0.19.3"
geopandas = "^0.14.0"
pyarrow = "^13.0.0"
ipython = "^8.15.0"
lxml = "^4.9.3"
tqdm = "^4.66.1"
more-itertools = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
